+++ models/reading.go
@@ -4,8 +4,8 @@ import "time"
 
 type Reading struct {
 	UserID     int       `json:"user_id"`
-	Current    int       `json:"current"`
-	Power      int       `json:"power"`
-	TotalPower int       `json:"total_power" binding:"required"`
+	Current    float32   `json:"current"`
+	Power      float32   `json:"power"`
+	TotalPower float32   `json:"total_power" binding:"required"`
 	Time       time.Time `json:"time"`
 }
